third_party/WebKit/LayoutTests/inspector/tracing/compile-script-expected.txt

CONSOLE MESSAGE: line 1: SCRIPT TAG
Tests the Timeline instrumentation for CompileScript event.

v8.compile Properties:
{
    data : {
        columnNumber : <number>
        lineNumber : <number>
        url : 
    }
    endTime : <number>
    startTime : <number>
    type : "v8.compile"
}
Text details for v8.compile: undefined
v8.compile Properties:
{
    data : {
        columnNumber : <number>
        lineNumber : <number>
        url : .../inspector/tracing/timeline-script-tag-2.js
    }
    endTime : <number>
    startTime : <number>
    type : "v8.compile"
}
Text details for v8.compile: timeline-script-tag-2.js:1
